Stop providing links for invalid paths
Gitlab can't (or doesn't, for security reasons) tell the difference
between a non-existent or private repository. So, if an invalid Gitlab
ticket path is provided, it will return the path along with a ticket
title for the sign in page. This change raises an IndexError if the path
is invalid.